refactor: useFilterData hook 생성
