+++ src/components/search/SearchBox.tsx
@@ -1,19 +1,14 @@
 import { useState, useEffect, useRef } from "react";
 import Select from "react-select";
 import SearchList from "./SearchList";
-import {
-  useGetCourseQuery,
-  useGetCourseConditionallyQuery,
-} from "../../redux/modules/apiSlice";
 import { hashTagOptions } from "../post/PostHashTag";
 import { regionOptions } from "../post/PostCategories";
-import { logEvent } from "../../utils/amplitude";
 import Skeleton, { SkeletonTheme } from "react-loading-skeleton";
 import { useNavigate } from "react-router";
 import { useSearchParams } from "react-router-dom";
 import { MultiStyles, SingleStyles } from "../shared";
 import { GrPowerReset } from "react-icons/gr";
-import { useOption } from "../../hooks";
+import { useOption, useFilterData } from "../../hooks";
 
 const travelStatusOptions: any = [
   { value: false, label: "여행 계획" },
@@ -23,11 +18,8 @@ const travelStatusOptions: any = [
 const SearchBox = () => {
   const navigate = useNavigate();
   const [hashtags2, sethashtags2] = useState<optionType[]>([]);
-  const [words, setWords] = useState("");
-  const [travelStatus, setTravelStatus] = useState<optionType | undefined>();
-  const [filteredList, setFilteredList] = useState<CourseType[]>();
+
   const [searchParams] = useSearchParams();
-  const { data } = useGetCourseQuery();
 
   const {
     selectedTags,
@@ -38,6 +30,18 @@ const SearchBox = () => {
     selectedRegions,
   } = useOption();
 
+  const {
+    filterData,
+    setTravelStatus,
+    setWords,
+    travelStatus,
+    words,
+    filteredList,
+    data,
+    isLoading,
+    isError,
+  } = useFilterData(selectedLabels, selectedRegions);
+
   const selectInputRef = useRef<any>(null);
 
   const onClearSelect = () => {
@@ -45,14 +49,6 @@ const SearchBox = () => {
       selectInputRef.current.clearValue();
     }
   };
-
-  const {
-    data: conditionData,
-    isLoading,
-    isError,
-  } = useGetCourseConditionallyQuery(travelStatus ? travelStatus.value : "");
-
-  //섹렉트된 데이터 형태 object에서 string[]으로 바꾸기
 
   //셀렉트한 데이터 State에 반영하기
   const locationOnChangeHandler = (data: any) => {
@@ -65,57 +61,6 @@ const SearchBox = () => {
 
   const travelStatusOnChangeHandler = (data: any) => {
     setTravelStatus(data);
-  };
-
-  //sample 배열이 base배열의 부분 함수인지 여부 true, false로 반환하는 함수
-  const isSubsetOf = function (
-    base: string[] | undefined,
-    sample: string[] | undefined
-  ) {
-    return sample?.every((item: any) => base?.includes(item));
-  };
-
-  // 선택된 검색 옵션이 없으면 전체 데이터 보여주기
-  const filterData = () => {
-    if (
-      words.length === 0 &&
-      selectedRegions?.length === 0 &&
-      selectedLabels?.length === 0 &&
-      travelStatus === undefined
-    ) {
-      setFilteredList(data);
-    }
-    // 지역, 해시태그, 키워드(제목, 코스 각각의 이름, 지번주소, 도로명 주소, 메모)에 따라 필터링
-    // 여행 여부 필터링은 firebase에서 해서 가져온다.
-    else {
-      const filteredData: CourseType[] | undefined = (
-        travelStatus !== undefined ? conditionData : data
-      )
-        ?.filter((item) => isSubsetOf(item.location, selectedRegions))
-        .filter((item) => isSubsetOf(item.hashtags, selectedLabels))
-        .filter(
-          (item) =>
-            item.nickname?.toLowerCase().includes(words.toLowerCase()) ||
-            item.title?.toLowerCase().includes(words.toLowerCase()) ||
-            JSON.parse(item.courseList).filter(
-              (item: CourseListType) =>
-                item.name.toLowerCase().includes(words.toLowerCase()) ||
-                item.address.toLowerCase().includes(words.toLowerCase()) ||
-                item.road.toLowerCase().includes(words.toLowerCase()) ||
-                item.memo?.toLowerCase().includes(words.toLowerCase())
-            ).length !== 0
-        );
-      setFilteredList(filteredData);
-    }
-    logEvent("게시물 검색", {
-      from: "검색페이지",
-      filter: {
-        해시태그: selectedLabels,
-        지역: selectedRegions,
-        여행여부: travelStatus,
-        검색어: words,
-      },
-    });
   };
 
   //맨 처음 렌더링, 새로고침 할 때 전체 데이터 보여주기
